refactor: move test and debug scripts to tests
