+++ tests/README.md
@@ -1,0 +1,13 @@
+# Test Utilities
+
+Legacy `test_*.php`, `debug_*.php` and related scripts are stored here.
+
+## Local test server
+
+To run these scripts locally, start the PHP development server:
+
+```bash
+./run_server.sh
+```
+
+Then open `http://localhost:8000/tests/` in your browser and choose the desired script.
