+++ tests/test_rename.php
@@ -1,0 +1,62 @@
+<?php
+chdir(dirname(__DIR__));
+/**
+ * VERIFICACIÓN FINAL - SISTEMA DE PERFILES RENOMBRADO
+ * Confirma que profile.php está funcionando correctamente
+ */
+
+echo "<h2>🔍 VERIFICACIÓN FINAL - PROFILE.PHP</h2>";
+echo "<hr>";
+
+// 1. Verificar que el archivo existe
+if (file_exists('profile.php')) {
+    echo "<div style='color: green;'>✅ Archivo profile.php existe</div>";
+} else {
+    echo "<div style='color: red;'>❌ Archivo profile.php NO encontrado</div>";
+}
+
+// 2. Verificar que perfil.php ya no existe
+if (!file_exists('perfil.php')) {
+    echo "<div style='color: green;'>✅ Archivo perfil.php ha sido eliminado correctamente</div>";
+} else {
+    echo "<div style='color: orange;'>⚠️ Archivo perfil.php aún existe (debería ser eliminado)</div>";
+}
+
+// 3. Verificar enlaces en navegación
+$companies_index = 'companies/index.php';
+if (file_exists($companies_index)) {
+    $content = file_get_contents($companies_index);
+    if (strpos($content, '../profile.php') !== false) {
+        echo "<div style='color: green;'>✅ Enlace a profile.php encontrado en companies/index.php</div>";
+    } else {
+        echo "<div style='color: orange;'>⚠️ Enlace a profile.php NO encontrado en companies/index.php</div>";
+    }
+    
+    if (strpos($content, '../perfil.php') !== false) {
+        echo "<div style='color: orange;'>⚠️ Aún hay referencias a perfil.php en companies/index.php</div>";
+    } else {
+        echo "<div style='color: green;'>✅ No hay referencias obsoletas a perfil.php</div>";
+    }
+} else {
+    echo "<div style='color: orange;'>⚠️ No se encontró companies/index.php para verificar</div>";
+}
+
+echo "<br>";
+echo "<h3>📋 RESUMEN:</h3>";
+echo "<ul>";
+echo "<li><strong>Archivo renombrado:</strong> perfil.php → profile.php</li>";
+echo "<li><strong>URL nueva:</strong> <a href='profile.php' target='_blank'>https://app.indiceapp.com/profile.php</a></li>";
+echo "<li><strong>Acceso desde dashboard:</strong> Menú usuario → Mi Perfil</li>";
+echo "</ul>";
+
+echo "<h3>🚀 PRÓXIMOS PASOS:</h3>";
+echo "<ol>";
+echo "<li>Probar el acceso a <strong>profile.php</strong> desde el navegador</li>";
+echo "<li>Verificar que el formulario guarde los datos correctamente</li>";
+echo "<li>Confirmar que el enlace del menú funcione</li>";
+echo "</ol>";
+
+echo "<hr>";
+echo "<div style='color: green; font-weight: bold;'>✅ SISTEMA DE PERFILES LISTO</div>";
+echo "<p>El archivo ha sido renombrado exitosamente y está listo para usar.</p>";
+?>
